Refactor api components from main into a package
